% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{get_cerebrovasc_disease_v1_codes}
\alias{get_cerebrovasc_disease_v1_codes}
\title{Retrieve ICD codes for cerebrovascular disease}
\usage{
get_cerebrovasc_disease_v1_codes(...)
}
\arguments{
\item{component}{Not used (fixed to \code{"cerebrovasc_disease_v1"}); included
only because it is inherited from \code{\link[=get_ascvd_codes]{get_ascvd_codes()}}.}
}
\description{
Direct access to the \code{cerebrovasc_disease_v1} component of \code{spec_ascvd}.
Equivalent to \code{get_ascvd_codes(component = "cerebrovasc_disease_v1", ...)}.
}
\examples{
get_cerebrovasc_disease_v1_codes()
}
\seealso{
\code{\link[=get_cerebrovasc_disease_v1_defs]{get_cerebrovasc_disease_v1_defs()}}, \code{\link[=get_ascvd_codes]{get_ascvd_codes()}},
\code{spec_ascvd}
}
