% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{get_chd_v1_defs}
\alias{get_chd_v1_defs}
\title{Retrieve the narrative algorithm description for coronary heart disease (CHD)}
\usage{
get_chd_v1_defs(...)
}
\arguments{
\item{component}{Not used (fixed to \code{"chd_v1"}); included only because it
is inherited from \code{\link[=get_ascvd_defs]{get_ascvd_defs()}}.}
}
\description{
Direct access to the \code{chd_v1} component of \code{spec_ascvd}. Equivalent to
\code{get_ascvd_defs(component = "chd_v1", ...)}.
}
\seealso{
\code{\link[=get_chd_v1_codes]{get_chd_v1_codes()}}, \code{\link[=get_ascvd_defs]{get_ascvd_defs()}}, \code{spec_ascvd}
}
