% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{get_hf_v1_defs}
\alias{get_hf_v1_defs}
\title{Retrieve the narrative algorithm description for heart failure (v1)}
\usage{
get_hf_v1_defs(variable_type = c("condition", "outcome"))
}
\arguments{
\item{variable_type}{\code{"condition"} (default) or \code{"outcome"}.}
}
\value{
Character string, or \code{NULL}.
}
\description{
Retrieve the narrative algorithm description for heart failure (v1)
}
\seealso{
\code{\link[=get_hf_v1_codes]{get_hf_v1_codes()}}
}
