% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{get_htn_v1_defs}
\alias{get_htn_v1_defs}
\alias{get_htn_v2_defs}
\title{Retrieve the narrative algorithm description for hypertension (v1)}
\usage{
get_htn_v1_defs(variable_type = c("condition", "outcome"), component = NULL)

get_htn_v2_defs(variable_type = c("condition", "outcome"), component = NULL)
}
\arguments{
\item{variable_type}{\code{"condition"} (default) or \code{"outcome"}.}

\item{component}{Not used. Pass \code{NULL} (default).}
}
\value{
Character string, or \code{NULL}.
}
\description{
Retrieve the narrative algorithm description for hypertension (v1)
}
\examples{
get_htn_v1_defs()
}
\seealso{
\code{\link[=get_htn_v1_codes]{get_htn_v1_codes()}}

\code{\link[=get_htn_v2_codes]{get_htn_v2_codes()}}
}
