% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{get_lead_pad_v1_codes}
\alias{get_lead_pad_v1_codes}
\title{Retrieve ICD codes for lower extremity artery disease (LEAD) / peripheral artery disease (PAD)}
\usage{
get_lead_pad_v1_codes(...)
}
\arguments{
\item{component}{Not used (fixed to \code{"lead_pad_v1"}); included only
because it is inherited from \code{\link[=get_ascvd_codes]{get_ascvd_codes()}}.}
}
\description{
Direct access to the \code{lead_pad_v1} component of \code{spec_ascvd}. Equivalent
to \code{get_ascvd_codes(component = "lead_pad_v1", ...)}.
}
\examples{
get_lead_pad_v1_codes()
}
\seealso{
\code{\link[=get_lead_pad_v1_defs]{get_lead_pad_v1_defs()}}, \code{\link[=get_ascvd_codes]{get_ascvd_codes()}}, \code{spec_ascvd}
}
