% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{get_ascvd_defs}
\alias{get_ascvd_defs}
\title{Retrieve the narrative algorithm description for an ASCVD component}
\usage{
get_ascvd_defs(variable_type = c("condition", "outcome"), component)
}
\arguments{
\item{variable_type}{\code{"condition"} (default) or \code{"outcome"}.}

\item{component}{\strong{Required.} Component name. See \code{\link[=get_ascvd_codes]{get_ascvd_codes()}}.}
}
\description{
Retrieve the narrative algorithm description for an ASCVD component
}
\seealso{
\code{\link[=get_ascvd_codes]{get_ascvd_codes()}}, \code{spec_ascvd}
}
