% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_spec_api.R
\name{modify_drug_spec}
\alias{modify_drug_spec}
\title{Modify a drug class specification}
\usage{
modify_drug_spec(
  spec,
  label = NULL,
  defs = NULL,
  generic_names = NULL,
  ndc = NULL
)
}
\arguments{
\item{spec}{A \link{DrugSpec} object.}

\item{label}{Optional replacement label string.}

\item{defs}{Optional replacement narrative string.}

\item{generic_names}{Optional replacement GNN character vector.}

\item{ndc}{Optional replacement NDC character vector.}
}
\value{
A modified deep clone of \code{spec}.
}
\description{
Returns a deep clone of \code{spec} with the requested modifications applied.
}
